Fix useEffect in NFT page
