+++ src/utils/data-transform-utils.js
@@ -22,3 +22,15 @@ export const transformContributorsData = (
       contributionToken.symbol,
   }
 }
+
+export const generateTimeDisplay = endDate => {
+  const rightJustNow = new Date().getTime()
+  const runway = endDate - rightJustNow
+  const stateObj = {
+    days: Math.floor(runway / (1000 * 60 * 60 * 24)),
+    hours: Math.floor((runway % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60)),
+    minutes: Math.floor((runway % (1000 * 60 * 60)) / (1000 * 60)),
+    seconds: Math.floor((runway % (1000 * 60)) / 1000),
+  }
+  return stateObj
+}
